+++ component/auth/wechatlogin.json
@@ -1,0 +1,4 @@
+{
+  "component": true,
+  "usingComponents": {}
+}
