+++ pages/register/register.json
@@ -1,1 +1,5 @@
-{}
+{
+  "usingComponents": {
+    "wechatlogin": "/component/auth/wechatlogin"
+  }
+}
